% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_laplacian_matrix_with_gausskernel.R
\name{calculate_laplacian_matrix_with_gausskernel}
\alias{calculate_laplacian_matrix_with_gausskernel}
\title{calculate_laplacian_matrix_with_gausskernel}
\usage{
calculate_laplacian_matrix_with_gausskernel(spa_mat, sigma = 1)
}
\arguments{
\item{spa_mat}{spa_mat}

\item{sigma}{sigma}
}
\description{
calculate_laplacian_matrix_with_gausskernel
}
