% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellChatDB.human}
\alias{CellChatDB.human}
\title{Ligand–receptor interaction database for human (CellChat)}
\format{
A list includes the ligand-receptor interactions
}
\source{
\url{https://github.com/sqjin/CellChat/}
}
\usage{
CellChatDB.human
}
\description{
A curated database of human ligand–receptor interactions used by the CellChat package to infer and analyze cell–cell communication networks from single-cell transcriptomic data.
}
\keyword{datasets}
