% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_good_weight.R
\name{define_good_weight}
\alias{define_good_weight}
\title{define_good_weight}
\usage{
define_good_weight(X, Wzero, Hzero, theta_v, weight_input)
}
\arguments{
\item{X}{X}

\item{Wzero}{Wzero}

\item{Hzero}{Hzero}

\item{theta_v}{theta_v}

\item{weight_input}{weight_input}
}
\description{
define_good_weight
}
