% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_2d.R
\name{distribution_2d}
\alias{distribution_2d}
\title{distribution_2d}
\usage{
distribution_2d(
  feature_matrix,
  one_feature,
  coordinate,
  thre_to_zero,
  color_palette = c("lightgrey", "blue"),
  visualization_type = "point",
  pt.size = 1,
  axis.title.size = 18,
  axis.text.size = 12,
  plot.title.size = 14,
  raster = F
)
}
\arguments{
\item{feature_matrix}{A matrix representing features, which could be topics, interactions, or gene expressions. The matrix should have features as its rows and samples/observations as its columns.}

\item{one_feature}{A single feature name from the feature_matrix that the user wants to visualize.}

\item{coordinate}{A data frame or matrix ideally sourced from a Seurat object. It should contain three columns: "row", "col", and "SB".}

\item{thre_to_zero}{A threshold value used to truncate low expression values.}

\item{color_palette}{A character vector specifying the color scheme for visualization. The default colors are light grey and blue. If given a single value, it should be one of the following: "viridis", "magma", "cividis", "inferno", "plasma". If given two values, they specify the gradient colors for low and high expression values, respectively.}

\item{visualization_type}{A character string indicating the type of visualization. The default is "point", but "density" can also be chosen to visualize data density.}

\item{pt.size}{A numeric value determining the size of the points in the plot.}

\item{axis.title.size}{A numeric value determining the font size of the axis titles.}

\item{axis.text.size}{A numeric value determining the font size of the axis text.}

\item{plot.title.size}{A numeric value determining the font size of the plot's title.}

\item{raster}{A boolean value indicating whether to rasterize the output plot. This can be useful for plots with a large number of points to improve rendering performance. By default, it's set to FALSE.}
}
\value{
a ggplot2 object
}
\description{
distribution_2d
}
