% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spot_by_celltype_extended.R
\name{get_spot_by_celltype_extended}
\alias{get_spot_by_celltype_extended}
\title{get_spot_by_celltype_extended}
\usage{
get_spot_by_celltype_extended(
  spatialdf,
  spot_by_celltype,
  smoothing_type = c("loop0", "loop1", "loop1+loop2"),
  distance_thre
)
}
\arguments{
\item{spatialdf}{a data frame or matrix that contains two columns, "col" and "row". The row names represent the barcode of the spot.}

\item{spot_by_celltype}{a data frame or matrix representing the abundance of cell types, with spots as rows and cell types as columns. Its row names should be consistent with the row names of 'spatialdf'.}

\item{smoothing_type}{This parameter has three options: c('loop0'), c('loop0', 'loop1'), and c('loop0', 'loop1', 'loop1+loop2'). 'loop0', 'loop1', and 'loop1+loop2' indicate the range within which data smoothing is performed, denoting: no smoothing, smoothing using one adjacent spot, and smoothing using two surrounding spots, respectively. As it involves the merging of matrices, the returned matrix's dimension varies depending on the chosen option. The dimensions corresponding to the three options are 1x, 2x, and 3x of the original dimensions, respectively.}

\item{distance_thre}{This parameter represents distance, defining the central distance from loop0 to loop1, and from loop0 to loop2. The length of the vector is consistent with the smoothing_type parameter. For example, distance_thre = c(NA,round(2/(3^0.5),2),round(4/(3^0.5),2))}
}
\description{
get_spot_by_celltype_extended
}
