% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topic_lr.R
\name{plot_topic_lr}
\alias{plot_topic_lr}
\title{Plot topic-specific ligand–receptor (LR) interaction heatmap}
\usage{
plot_topic_lr(
  topic_lr,
  topic_lr_quantile_thre = 0.95,
  top_num_plot = 10,
  fill_viridis = "F",
  axis.title.size = 14,
  axis.text.x.bottom.size = 12,
  axis.text.y.left.size = 8,
  strip.text.y.size = 12
)
}
\arguments{
\item{topic_lr}{A numeric data frame or matrix with LR interactions as rows
and topics as columns. Row names must correspond to LR interaction names.}

\item{topic_lr_quantile_thre}{A numeric value between 0 and 1 specifying the
global quantile threshold used to filter LR values across all topics.
Only LR–topic pairs with values above this quantile will be considered.
Default is \code{0.95}.}

\item{top_num_plot}{The maximum number of LR interactions to be displayed for each topic in the plot. Default is 10.}

\item{fill_viridis}{Specifies the "viridis" color palette option to be used for the heatmap. Default is "F".}

\item{axis.title.size}{The font size of the axis titles. Default is 14.}

\item{axis.text.x.bottom.size}{The font size for the text on the bottom of the x-axis. Default is 12.}

\item{axis.text.y.left.size}{The font size for the text on the left of the y-axis. Default is 8.}

\item{strip.text.y.size}{The font size for the text on the y-axis strip. Default is 12.}
}
\value{
A \code{ggplot2} object representing a faceted heatmap of topic-specific LR interactions.
}
\description{
This function visualizes topic-specific ligand–receptor (LR) interactions
as a heatmap. For each topic, LR values are first normalized to
relative contributions, then filtered by a global quantile threshold.
The top LR interactions for each topic are selected and displayed
in separate facets.
}
