% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_LRintegratedmatrix.R
\name{prepare_for_LRintegratedmatrix}
\alias{prepare_for_LRintegratedmatrix}
\title{prepare_for_LRintegratedmatrix}
\usage{
prepare_for_LRintegratedmatrix(seu_obj, ccc_db)
}
\arguments{
\item{seu_obj}{A Seurat object containing spatial transcriptomics data, with 'Spatial' as the assay name. The object should be normalized using the NormalizeData function.}

\item{ccc_db}{This parameter represents the reference database of cell-cell interactions, e.g., CellChatDB.human, containing interaction, complex, and geneInfo tables.}
}
\value{
A list with two elements: (1) a gene/complex-by-spot expression matrix, and (2) a filtered CellChat interaction data frame.
}
\description{
This function prepares an integrated expression matrix for ligand–receptor analysis in spatial transcriptomics data.
It filters CellChat ligand–receptor interactions based on genes present in the Spatial assay of a Seurat object and returns both the integrated expression matrix and the filtered interaction table.
}
