% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaNiche-package.R
\docType{package}
\name{SpaNiche-package}
\alias{SpaNiche}
\alias{SpaNiche-package}
\title{SpaNiche: Spatial Niche Analysis}
\description{
Provides tools for spatial niche analysis to explore colocalization patterns and cellular interactions in spatial transcriptomics data.
}
\author{
\strong{Maintainer}: Siyuan Huang \email{huangsiyuan1001@163.com}

}
\keyword{internal}
