% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatme-package.R
\docType{package}
\name{spatme-package}
\alias{spatme}
\alias{spatme-package}
\title{spatme: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Siyuan Huang \email{huangsiyuan1001@163.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
