% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatme_nmf.R
\name{spatme_nmf}
\alias{spatme_nmf}
\title{spatme_nmf}
\usage{
spatme_nmf(
  vis.seu,
  spatialdf,
  enrichmentdf,
  smoothing_type = c("loop0", "loop1", "loop1+loop2"),
  distance_thre = c(NA, round(2/(3^0.5), 2), round(4/(3^0.5), 2)),
  LRDB = spatme::CellChatDB.human,
  LR_dm_type = c("loop1", "loop2"),
  LR_distance_thre = c(round(2/(3^0.5), 2), round(4/(3^0.5), 2)),
  var_thre = c(0.2, 0.99),
  topic_num = 15,
  defined_weight = "default",
  lambda_v = c(0.5, 1, 2),
  sigma_v = c(0.5, 1, 1.5),
  maxiter = 200,
  st.count = 20,
  epsilon = 1e-04
)
}
\arguments{
\item{vis.seu}{ST data is processed as a Seurat object with 'Spatial' as the assay name. Seurat object is normalized by using the 'NormalizeData' function.}

\item{spatialdf}{a data frame or matrix that contains two columns, "col" and "row". The row names represent the barcode of the spot.}

\item{enrichmentdf}{a data frame or matrix representing the abundance of cell types, with spots as rows and cell types as columns. Its row names should be consistent with the row names of 'spatialdf'.}

\item{smoothing_type}{This parameter has three options: c('loop0'), c('loop0', 'loop1'), and c('loop0', 'loop1', 'loop1+loop2'). 'loop0', 'loop1', and 'loop1+loop2' indicate the range within which data smoothing is performed, denoting: no smoothing, smoothing using one adjacent spot, and smoothing using two surrounding spots, respectively. As it involves the merging of matrices, the returned matrix's dimension varies depending on the chosen option. The dimensions corresponding to the three options are 1x, 2x, and 3x of the original dimensions, respectively.}

\item{distance_thre}{This parameter represents distance, defining the central distance from loop0 to loop1, and from loop0 to loop2. The length of the vector is consistent with the smoothing_type parameter. For example, distance_thre = c(NA,round(2/(3^0.5),2),round(4/(3^0.5),2))}

\item{LRDB}{This parameter represents the reference for cell-cell interactions, which comes from CellChat.}

\item{LR_dm_type}{This parameter can be one or more values from c('loop1', 'loop2'), representing the type of adjacency matrix to be returned.}

\item{LR_distance_thre}{This parameter represents distance, defining the central distance from loop0 to loop1(, and from loop0 to loop2).}

\item{var_thre}{Variance threshold}

\item{topic_num}{An integer specifying the number of components or topics to be extracted. It determines the number of columns in the W matrix and the number of rows in each H matrix.}

\item{defined_weight}{defined_weight}

\item{lambda_v}{lambda}

\item{sigma_v}{Gaussian kernel parameter. Used in the computation of the Laplacian matrix, influencing how spatial information is incorporated into the factorization.}

\item{maxiter}{The maximum number of iterations allowed for the factorization process. This acts as a stopping criterion to prevent the algorithm from running indefinitely.}

\item{st.count}{Convergence counter. If the change in the reconstruction error remains below epsilon for st.count consecutive iterations, the algorithm stops, assuming it has converged.}

\item{epsilon}{A threshold for the change in reconstruction error. If the relative change in error is less than epsilon for st.count consecutive iterations, convergence is assumed.}
}
\description{
spatme_nmf
}
