% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_for_interactionwithmoransi.R
\name{stat_for_interactionwithmoransi}
\alias{stat_for_interactionwithmoransi}
\title{stat_for_interactionwithmoransi}
\usage{
stat_for_interactionwithmoransi(
  moransi_df,
  moransi_thre = 0.2,
  p.value_thre = 0.05,
  height_ratio = c(5, 2)
)
}
\arguments{
\item{moransi_df}{A data frame expected to contain columns such as MoransI_observed, MoransI_p.value, pathway_name, and annotation.}

\item{moransi_thre}{Threshold for the Moran's I value to filter the interactions.}

\item{p.value_thre}{Threshold for the p-value to filter the interactions.}

\item{height_ratio}{Ratio to define the relative heights of the two plots when they are combined.}
}
\description{
stat_for_interactionwithmoransi
}
