% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_initial_weight.R
\name{define_initial_weight}
\alias{define_initial_weight}
\title{define_initial_weight}
\usage{
define_initial_weight(data.list)
}
\arguments{
\item{data.list}{data.list}
}
\description{
define_initial_weight
}
