% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_LRintegratedmatrix.R
\name{generate_LRintegratedmatrix}
\alias{generate_LRintegratedmatrix}
\title{generate_LRintegratedmatrix}
\usage{
generate_LRintegratedmatrix(
  neighbor_matrix,
  neighbor_role,
  center_role,
  center_is_ligand
)
}
\arguments{
\item{neighbor_matrix}{A matrix representing the relationship between neighboring spatial locations.}

\item{neighbor_role}{A matrix (spot by gene).}

\item{center_role}{A matrix (spot by gene).}

\item{center_is_ligand}{A logical variable indicating if the central role represents a ligand. If TRUE, the central role is a ligand; if FALSE, it is a receptor.}
}
\description{
generate_LRintegratedmatrix
}
