% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_distance_matrix.R
\name{make_distance_matrix}
\alias{make_distance_matrix}
\title{make_distance_matrix}
\usage{
make_distance_matrix(
  spatialdf,
  dm_type = c("view1", "view2"),
  distance_thre,
  digits = 2
)
}
\arguments{
\item{spatialdf}{a data frame or matrix that contains two columns, "col" and "row". Row names should  correspond to spot barcodes.}

\item{dm_type}{This parameter can be one or more values from c('view1', 'view2'), representing the type of adjacency matrix to be returned. For example, you can choose c('view1') or c('view1', 'view2').}

\item{distance_thre}{This parameter represents distance, defining the central distance from center to it's first/second-order neighbors. Consistent with dm_type, you can define c(c1) or c(c1, c2), and c1/c2 is distance value.}

\item{digits}{integer indicating the number of decimal places.}
}
\value{
A named list of adjacency matrices corresponding to each dm_type.
Each matrix is normalized so that rows sum to 1 (row-normalization matrix).
}
\description{
The function calculates Euclidean distances between spatial spots and converts them into adjacency matrices based on specified distance thresholds.
}
