% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_distance_matrix_LR.R
\name{make_distance_matrix_LR}
\alias{make_distance_matrix_LR}
\title{make_distance_matrix_LR}
\usage{
make_distance_matrix_LR(
  spatialdf,
  dm_type = c("loop1", "loop2"),
  distance_thre
)
}
\arguments{
\item{spatialdf}{a data frame or matrix that contains two columns, "col" and "row". The row names represent the barcode of the spot.}

\item{dm_type}{This parameter can be one or more values from c('loop1', 'loop2'), representing the type of adjacency matrix to be returned.}

\item{distance_thre}{This parameter represents distance, defining the central distance from loop0 to loop1(, and from loop0 to loop2).}
}
\description{
make_distance_matrix_LR
}
