% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moransi_for_LRintegratedmatrix.R
\name{moransi_for_LRintegratedmatrix}
\alias{moransi_for_LRintegratedmatrix}
\title{moransi_for_LRintegratedmatrix}
\usage{
moransi_for_LRintegratedmatrix(LRintegratedmatrix, spatial.seu, cc_interaction)
}
\arguments{
\item{LRintegratedmatrix}{integrated LR matrix, where the row names represent features and the column names represent spots. The underscore ('_') in the row names will be replaced by a hyphen ('-').}

\item{spatial.seu}{spatial.seu}

\item{cc_interaction}{cc_interaction}
}
\description{
moransi_for_LRintegratedmatrix
}
