% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NMF_components.R
\name{plot_NMF_components}
\alias{plot_NMF_components}
\title{plot_NMF_components}
\usage{
plot_NMF_components(
  components,
  one_topic = "topic1",
  p1.axis.text.x.bottom.size = 8,
  p2.axis.text.x.bottom.size = 9,
  p1.strip.text.x.size = 15,
  p2.strip.text.x.size = 9,
  color_celltype = NULL,
  color_loop = NULL
)
}
\arguments{
\item{components}{A dataframe containing the NMF component data.}

\item{one_topic}{A string that specifies which topic to visualize. Default is "topic1".}

\item{p1.axis.text.x.bottom.size}{p1.axis.text.x.bottom.size}

\item{p2.axis.text.x.bottom.size}{p2.axis.text.x.bottom.size}

\item{p1.strip.text.x.size}{p1.strip.text.x.size}

\item{p2.strip.text.x.size}{p2.strip.text.x.size}

\item{color_celltype}{Either NULL or a vector of colors. If provided, it specifies the colors to use for the different cell types in the plot. If NULL, default ggplot2 colors are used.}

\item{color_loop}{Either NULL or a vector of colors. If provided, it specifies the colors to use for the different loops in the plot. If NULL, default ggplot2 colors are used.}
}
\value{
A combined ggplot2 plot that visualizes the values of the specified topic grouped by celltype and loop.
}
\description{
plot_NMF_components
}
