% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topic_lr.R
\name{plot_topic_lr}
\alias{plot_topic_lr}
\title{plot_topic_lr}
\usage{
plot_topic_lr(
  topic_lr,
  topic_lr_quantile_thre = 0.95,
  top_num_plot = 10,
  fill_viridis = "F",
  axis.title.size = 14,
  axis.text.x.bottom.size = 12,
  axis.text.y.left.size = 8,
  strip.text.y.size = 12
)
}
\arguments{
\item{topic_lr}{A dataframe (LR by topic) representing the relationship between topics and LR.}

\item{topic_lr_quantile_thre}{A threshold for quantile selection. Only data above this quantile will be considered for visualization. Default is 0.95.}

\item{top_num_plot}{The maximum number of LR interactions to be displayed for each topic in the plot. Default is 10.}

\item{fill_viridis}{Specifies the "viridis" color palette option to be used for the heatmap. Default is "F".}

\item{axis.title.size}{The font size of the axis titles. Default is 14.}

\item{axis.text.x.bottom.size}{The font size for the text on the bottom of the x-axis. Default is 12.}

\item{axis.text.y.left.size}{The font size for the text on the left of the y-axis. Default is 8.}

\item{strip.text.y.size}{The font size for the text on the y-axis strip. Default is 12.}
}
\description{
plot_topic_lr
}
