% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_LRintegratedmatrix.R
\name{prepare_for_LRintegratedmatrix}
\alias{prepare_for_LRintegratedmatrix}
\title{prepare_for_LRintegratedmatrix}
\usage{
prepare_for_LRintegratedmatrix(seu_obj, ccc_db)
}
\arguments{
\item{seu_obj}{ST data is processed as a Seurat object with 'Spatial' as the assay name. Seurat object is normalized by using the 'NormalizeData' function.}

\item{ccc_db}{This parameter represents the reference for cell-cell interactions, which comes from CellChat.}
}
\description{
prepare_for_LRintegratedmatrix
}
